rename share to qabeluser to forward
