+++ qabelbox/src/main/java/de/qabel/qabelbox/storage/AbstractNavigation.java
@@ -5,6 +5,29 @@ import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.util.Log;
 
+import org.apache.commons.io.IOUtils;
+import org.apache.commons.lang3.NotImplementedException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.spongycastle.crypto.params.KeyParameter;
+
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
+import java.security.InvalidKeyException;
+import java.util.AbstractMap.SimpleEntry;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
+import java.util.Stack;
+import java.util.UUID;
+
 import de.qabel.core.crypto.CryptoUtils;
 import de.qabel.core.crypto.QblECKeyPair;
 import de.qabel.core.crypto.QblECPublicKey;
@@ -12,17 +35,6 @@ import de.qabel.qabelbox.exceptions.QblStorageException;
 import de.qabel.qabelbox.exceptions.QblStorageNameConflict;
 import de.qabel.qabelbox.exceptions.QblStorageNotFound;
 import de.qabel.qabelbox.providers.BoxProvider;
-
-import org.apache.commons.io.IOUtils;
-import org.apache.commons.lang3.NotImplementedException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.spongycastle.crypto.params.KeyParameter;
-
-import java.io.*;
-import java.security.InvalidKeyException;
-import java.util.*;
-import java.util.AbstractMap.SimpleEntry;
 
 public abstract class AbstractNavigation implements BoxNavigation {
 
@@ -57,7 +69,7 @@ public abstract class AbstractNavigation implements BoxNavigation {
 		this.transferManager = transferManager;
 		this.boxVolume = boxVolume;
 		this.currentPath = path;
-        this.dmKey = dmKey;
+		this.dmKey = dmKey;
 		this.context = context;
 		this.cache = new FileCache(context);
 		cryptoUtils = new CryptoUtils();
@@ -74,7 +86,7 @@ public abstract class AbstractNavigation implements BoxNavigation {
 
 	public String getPath(BoxObject object) {
 		if (object instanceof BoxFolder) {
-			return currentPath+ object.name + BoxProvider.PATH_SEP;
+			return currentPath + object.name + BoxProvider.PATH_SEP;
 		} else {
 			return currentPath + object.name;
 		}
@@ -99,7 +111,7 @@ public abstract class AbstractNavigation implements BoxNavigation {
 		if (transferManager.waitFor(id)) {
 			return file;
 		} else {
-			throw new QblStorageNotFound("File not found. Prefix: " + prefix  + " Name: " + name);
+			throw new QblStorageNotFound("File not found. Prefix: " + prefix + " Name: " + name);
 		}
 	}
 
@@ -130,22 +142,23 @@ public abstract class AbstractNavigation implements BoxNavigation {
 	}
 
 	/**
-     * Navigates to a direct subfolder.
-     * @param target Subfolder to navigate to
-     * @throws QblStorageException
-     */
+	 * Navigates to a direct subfolder.
+	 *
+	 * @param target Subfolder to navigate to
+	 * @throws QblStorageException
+	 */
 	@Override
 	public void navigate(BoxFolder target) throws QblStorageException {
-        boolean isSubfolder = false;
-        for (BoxFolder boxFolder : listFolders()) {
-            if (boxFolder.ref.equals(target.ref)) {
-                isSubfolder = true;
-                break;
-            }
-        }
-        if (!isSubfolder) {
-            throw new QblStorageNotFound(target.name + " is not a direct subfolder of " + currentPath);
-        }
+		boolean isSubfolder = false;
+		for (BoxFolder boxFolder : listFolders()) {
+			if (boxFolder.ref.equals(target.ref)) {
+				isSubfolder = true;
+				break;
+			}
+		}
+		if (!isSubfolder) {
+			throw new QblStorageNotFound(target.name + " is not a direct subfolder of " + currentPath);
+		}
 		try {
 			doNavigate(target, true);
 		} catch (QblStorageException e) {
@@ -177,7 +190,7 @@ public abstract class AbstractNavigation implements BoxNavigation {
 					dmKey = target.key;
 				}
 			}
-		} catch(IOException | InvalidKeyException e){
+		} catch (IOException | InvalidKeyException e) {
 			throw new QblStorageException(e);
 		}
 	}
@@ -198,16 +211,16 @@ public abstract class AbstractNavigation implements BoxNavigation {
 			// ignore our local directory metadata
 			// all changes that are not inserted in the new dm are _lost_!
 			dm = updatedDM;
-			for (FileUpdate update: updatedFiles) {
+			for (FileUpdate update : updatedFiles) {
 				handleConflict(update);
 			}
 			dm.commit();
 		}
-		for (FileUpdate update: updatedFiles) {
+		for (FileUpdate update : updatedFiles) {
 			updateFileMetadata(update.updated);
 		}
 		uploadDirectoryMetadata();
-		for (String ref: deleteQueue) {
+		for (String ref : deleteQueue) {
 			blockingDelete(prefix, ref);
 		}
 		// TODO: make a test fail without these
@@ -285,8 +298,7 @@ public abstract class AbstractNavigation implements BoxNavigation {
 					FileMetadata fileMetadata = new FileMetadata(out);
 					boxExternals.add(fileMetadata.getFile());
 				}
-			}
-			catch (QblStorageException e) {
+			} catch (QblStorageException e) {
 				Log.e(TAG, "Cannot load metadata file: " + boxExternalRefs.getPrefix() + '/' + boxExternalRefs.getBlock());
 				if (boxExternalRefs.isFolder) {
 					boxExternals.add(new BoxExternalFolder(boxExternalRefs.url, boxExternalRefs.name,
@@ -307,7 +319,7 @@ public abstract class AbstractNavigation implements BoxNavigation {
 		File out = new File(context.getExternalCacheDir(), UUID.randomUUID().toString());
 
 		try (InputStream decryptedInputStream = openStream(key, encryptedMetadata);
-			 FileOutputStream fileOutputStream = new FileOutputStream(out)){
+			 FileOutputStream fileOutputStream = new FileOutputStream(out)) {
 			IOUtils.copy(decryptedInputStream, fileOutputStream);
 		} catch (IOException e) {
 			throw new QblStorageException("Could not decrypt FileMetadata", e);
@@ -377,38 +389,51 @@ public abstract class AbstractNavigation implements BoxNavigation {
 	 * and encryption key to BoxFile.metakey. If BoxFile.meta or BoxFile.metakey is not null, BoxFile will not be
 	 * modified and no FileMetadata will be created. Call {@link #removeFileMetadata(BoxFile boxFile)}
 	 * first if you want to re-create a new FileMetadata.
+	 *
 	 * @param boxFile BoxFile to create FileMetadata from.
 	 * @return True if FileMetadata has successfully created and uploaded.
 	 */
 	@Override
 	public BoxExternalReference createFileMetadata(QblECPublicKey owner, BoxFile boxFile) throws QblStorageException {
-		if (boxFile.meta != null || boxFile.metakey != null) {
-			return new BoxExternalReference(false, boxFile.prefix + '/' + boxFile.meta, boxFile.name, owner, boxFile.metakey);
-		}
-		String metaBlock = UUID.randomUUID().toString();
-		KeyParameter key = cryptoUtils.generateSymmetricKey();
-		boxFile.meta = metaBlock;
-		boxFile.metakey = key.getKey();
 
 		try {
-			FileMetadata fileMetadata = new FileMetadata(owner, boxFile, dm.getTempDir());
-			FileInputStream fileInputStream = new FileInputStream(fileMetadata.getPath());
-			uploadEncrypted(fileInputStream, key, prefix, metaBlock, null);
-
-			// Overwrite = delete old file, upload new file
-			BoxFile oldFile = dm.getFile(boxFile.name);
-			if (oldFile != null) {
-				dm.deleteFile(oldFile);
+			//check if file a received share
+			if (getMetadataFile(boxFile.prefix, boxFile.block, boxFile.key) != null) {
+				return new BoxExternalReference(false, boxFile.prefix + '/' + boxFile.meta, boxFile.name, owner, boxFile.metakey);
 			}
-			dm.insertFile(boxFile);
-		} catch (QblStorageException | FileNotFoundException e) {
-			throw new QblStorageException("Could not create or upload FileMetadata", e);
+		} catch (Exception ex) {
+			//no. create new
+			if (boxFile.meta != null || boxFile.metakey != null) {
+				return new BoxExternalReference(false, boxFile.prefix + '/' + boxFile.meta, boxFile.name, owner, boxFile.metakey);
+			}
+			String metaBlock = UUID.randomUUID().toString();
+			KeyParameter key = cryptoUtils.generateSymmetricKey();
+			boxFile.meta = metaBlock;
+			boxFile.metakey = key.getKey();
+
+			try {
+				FileMetadata fileMetadata = new FileMetadata(owner, boxFile, dm.getTempDir());
+				FileInputStream fileInputStream = new FileInputStream(fileMetadata.getPath());
+				uploadEncrypted(fileInputStream, key, prefix, metaBlock, null);
+
+				// Overwrite = delete old file, upload new file
+				BoxFile oldFile = dm.getFile(boxFile.name);
+				if (oldFile != null) {
+					dm.deleteFile(oldFile);
+				}
+				dm.insertFile(boxFile);
+			} catch (QblStorageException | FileNotFoundException e) {
+				throw new QblStorageException("Could not create or upload FileMetadata", e);
+			}
+			return new BoxExternalReference(false, boxFile.prefix + '/' + metaBlock, boxFile.name, owner, boxFile.metakey);
 		}
-		return new BoxExternalReference(false, boxFile.prefix + '/' + metaBlock, boxFile.name, owner, boxFile.metakey);
+		//how can this position reached?
+		return null;
 	}
 
 	/**
 	 * Updates and uploads a FileMetadata object for a BoxFile.
+	 *
 	 * @param boxFile BoxFile to create FileMetadata from.
 	 * @return True if FileMetadata has successfully created and uploaded.
 	 */
@@ -433,6 +458,7 @@ public abstract class AbstractNavigation implements BoxNavigation {
 	/**
 	 * Deletes FileMetadata and sets BoxFile.meta and BoxFile.metakey to null. Does not re-encrypt BoxFile thus
 	 * receivers of the FileMetadata can still read the BoxFile.
+	 *
 	 * @param boxFile BoxFile to remove FileMetadata from.
 	 * @return True if FileMetadata has been deleted. False if meta information is missing.
 	 */
@@ -451,16 +477,18 @@ public abstract class AbstractNavigation implements BoxNavigation {
 
 	/**
 	 * Attaches a received BoxExternal to the DirectoryMetadata
+	 *
 	 * @param boxExternalReference Reference to meta file
 	 * @throws QblStorageException If metadata cannot be accesses or decrypted.
 	 */
 	@Override
 	public void attachExternal(BoxExternalReference boxExternalReference) throws QblStorageException {
-			dm.insertExternalReference(boxExternalReference);
+		dm.insertExternalReference(boxExternalReference);
 	}
 
 	/**
 	 * Deletes a BoxExternalFile from the DirectoryMetadata.
+	 *
 	 * @param name Name of received shared BoxFile to delete from DirectoryMetadata.
 	 * @throws QblStorageException
 	 */
@@ -470,7 +498,7 @@ public abstract class AbstractNavigation implements BoxNavigation {
 	}
 
 	private File refreshCache(BoxFile boxFile, @Nullable TransferManager.BoxTransferListener boxTransferListener) throws QblStorageNotFound {
-		logger.info("Refreshing cache: "+ boxFile.block);
+		logger.info("Refreshing cache: " + boxFile.block);
 		File download = blockingDownload(boxFile.prefix, BLOCKS_PREFIX + boxFile.block, boxTransferListener);
 		cache.put(boxFile, download);
 		cache.close();
@@ -505,7 +533,7 @@ public abstract class AbstractNavigation implements BoxNavigation {
 		BoxFolder folder = new BoxFolder(dm.getFileName(), name, secretKey.getKey());
 		this.dm.insertFolder(folder);
 		BoxNavigation newFolder = new FolderNavigation(prefix, dm, keyPair, secretKey.getKey(),
-			deviceId, transferManager, boxVolume, currentPath + BoxProvider.PATH_SEP + folder.name,
+				deviceId, transferManager, boxVolume, currentPath + BoxProvider.PATH_SEP + folder.name,
 				parentBoxFolders, context);
 		newFolder.commit();
 		return folder;
@@ -531,14 +559,14 @@ public abstract class AbstractNavigation implements BoxNavigation {
 
 	@Override
 	public void delete(BoxFolder folder) throws QblStorageException {
-        navigate(folder);
-		for (BoxFile file: listFiles()) {
+		navigate(folder);
+		for (BoxFile file : listFiles()) {
 			logger.info("Deleting file " + file.name);
-            delete(file);
+			delete(file);
 		}
-		for (BoxFolder subFolder: listFolders()) {
+		for (BoxFolder subFolder : listFolders()) {
 			logger.info("Deleting folder " + folder.name);
-            delete(subFolder);
+			delete(subFolder);
 		}
 		navigateToParent();
 		commit();
